Avoid having to wrap `MaterialSelectDemoComponent`
This finally enables me to effeciently set up the test environment
without having to wrap the tested `Component` (as it was done so far).
`NgTestBed.setPageLoader(...)` is used now to set the `PageLoader`
parent element to a container which contains `NgTestBed`'s host element
and `overlayContainerParent`'s element.

At last, after 3 weeks of research, this finally works as I envisioned
it in the beginning - Soli Deo Gloria!!